requirements.txt
joblib==0.14.0
idna==2.8
configparser==3.7.4
psycopg2-binary==2.8.3
py==1.8.0
pytest==5.0.1
pytest-expect==1.1.0
pytest-html==1.21.1
requests==2.22.0
urllib3==1.25.3